requirements.txt
mysql-connector-python
psycopg2
pymongo
schedule
boto3
google-cloud-storage
azure-storage-blob
requests